requirements.txt

streamlit
pandas
python-docx
PyMuPDF
openpyxl
